requirements.txt
gitpython>=3.1.0
pyyaml>=6.0
configparser>=5.0.0
requests>=2.25.0